move 'converge' tests to separate file
