+++ types/ramda/tsconfig.json
@@ -52,6 +52,7 @@
         "test/cond-tests.ts",
         "test/construct-tests.ts",
         "test/constructN-tests.ts",
-        "test/contains-tests.ts"
+        "test/contains-tests.ts",
+        "test/converge-tests.ts"
     ]
 }
